+++ bec_widgets/cli/__init__.py
@@ -1,0 +1,1 @@
+from bec_widgets.cli.rpc import rpc_base
