+++ bec_widgets/utils/generate_cli.py
