+++ bec_widgets/widgets/plots/heatmap/settings/heatmap_setting.py
@@ -4,9 +4,9 @@ import os
 
 from qtpy.QtWidgets import QFrame, QScrollArea, QVBoxLayout
 
-from bec_widgets.utils import UILoader
 from bec_widgets.utils.error_popups import SafeSlot
 from bec_widgets.utils.settings_dialog import SettingWidget
+from bec_widgets.utils.ui_loader import UILoader
 
 
 class HeatmapSettings(SettingWidget):
