+++ bec_widgets/widgets/plots/multi_waveform/settings/control_panel.py
@@ -2,9 +2,9 @@ import os
 
 from qtpy.QtWidgets import QVBoxLayout, QWidget
 
-from bec_widgets.utils import UILoader
 from bec_widgets.utils.error_popups import SafeSlot
 from bec_widgets.utils.settings_dialog import SettingWidget
+from bec_widgets.utils.ui_loader import UILoader
 from bec_widgets.utils.widget_io import WidgetIO
