+++ pyproject.toml
@@ -12,19 +12,19 @@ dependencies = [
   "PyJWT~=2.9",
   "PySide6==6.9.0",
   "PySide6-QtAds==4.4.0",
-  "bec_ipython_client~=3.107,>=3.107.2",  # needed for jupyter console
+  "bec_ipython_client~=3.107,>=3.107.2", # needed for jupyter console
   "bec_lib~=3.107,>=3.107.2",
   "bec_qthemes~=1.0, >=1.3.4",
-  "black>=26,<27",  # needed for bw-generate-cli
+  "black>=26,<27",                       # needed for bw-generate-cli
   "copier~=9.7",
   "darkdetect~=0.8",
-  "isort>=5.13, <9.0",  # needed for bw-generate-cli
+  "isort>=5.13, <9.0",                   # needed for bw-generate-cli
   "markdown~=3.9",
   "ophyd_devices~=1.29, >=1.29.1",
   "pydantic~=2.0",
   "pylsp-bec~=1.2",
   "pyqtgraph==0.13.7",
-  "qtconsole~=5.5, >=5.5.1",  # needed for jupyter console
+  "qtconsole~=5.5, >=5.5.1",             # needed for jupyter console
   "qtmonaco~=0.8, >=0.8.1",
   "qtpy~=2.4",
   "thefuzz~=0.22",
@@ -38,8 +38,8 @@ Homepage = "https://gitlab.psi.ch/bec/bec_widgets"
 [project.scripts]
 bec-app = "bec_widgets.applications.main_app:main"
 bec-designer = "bec_widgets.utils.bec_designer:main"
-bec-gui-server = "bec_widgets.cli.server:main"
-bw-generate-cli = "bec_widgets.cli.generate_cli:main"
+bec-gui-server = "bec_widgets.applications.companion_app:main"
+bw-generate-cli = "bec_widgets.utils.generate_cli:main"
 
 [project.optional-dependencies]
 dev = [
@@ -55,9 +55,7 @@ dev = [
   "watchdog~=6.0",
   "pre_commit~=4.2",
 ]
-qtermwidget = [
-  "pyside6_qtermwidget",
-]
+qtermwidget = ["pyside6_qtermwidget"]
 
 [build-system]
 requires = ["hatchling"]
@@ -68,7 +66,7 @@ line-length = 100
 skip-magic-trailing-comma = true
 
 [tool.coverage.report]
-skip_empty = true  # exclude empty *files*, e.g. __init__.py, from the report
+skip_empty = true # exclude empty *files*, e.g. __init__.py, from the report
 exclude_lines = [
   "pragma: no cover",
   "if TYPE_CHECKING:",
