+++ tests/unit_tests/test_ring_progress_bar.py
@@ -8,7 +8,7 @@ from qtpy.QtCore import QEvent, QPoint, QPointF, Qt
 from qtpy.QtGui import QColor, QMouseEvent
 from qtpy.QtWidgets import QApplication
 
-from bec_widgets.utils import Colors
+from bec_widgets.utils.colors import Colors
 from bec_widgets.widgets.progress.ring_progress_bar.ring_progress_bar import (
     RingProgressBar,
     RingProgressContainerWidget,
